Rename host model
