+++ src/app/models/openlabs_range_model.py
@@ -3,7 +3,7 @@ from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from ..core.db.database import Base
 from ..enums.providers import OpenLabsProvider
-from .openlabs_base_model import OpenLabsTemplateMixin
+from .template_base_model import OpenLabsTemplateMixin
 
 
 class OpenLabsRangeModel(Base, OpenLabsTemplateMixin):
