+++ src/app/models/openlabs_vpc_model.py
@@ -6,7 +6,7 @@ from sqlalchemy.dialects.postgresql import CIDR, UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from ..core.db.database import Base
-from .openlabs_base_model import OpenLabsTemplateMixin
+from .template_base_model import OpenLabsTemplateMixin
 
 
 class OpenLabsVPCModel(Base, OpenLabsTemplateMixin):
